Cognitive Neural Networks
For this assignment I completed a total of 9 questions relating to some datasets we were given to analyse. Using this data we had to make python code that would initialise Hebb and Oja's rules to see patterns produced. A few python libraries were used for this, and a lot was learnt.